+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added `class` member to `api::v2::order::Order` type
+- Added `symbols` member to `api::v2::orders::OrdersReq` type
 - Bumped `uuid` dependency to `1.0`